factor out authentication to authenticatedPost
