+++ .travis.yml
@@ -8,3 +8,5 @@ before_script:
   - npm run migrations
 git:
     submodules: false
+env:
+    TIMESYNC_LOCAL=1
