+++ tests/projects.js
@@ -565,6 +565,7 @@ module.exports = function(expect, request, baseUrl) {
       auth: {
         username: 'tschuy',
         password: 'password',
+        type: 'password',
       },
       object: project,
     };
